Revert "Add tests for pinpoint node and browser"
