+++ packages/analytics-plugin-aws-pinpoint/package.json
@@ -21,9 +21,6 @@
   "author": "David Wells",
   "license": "MIT",
   "scripts": {
-    "test": "ava -v -s",
-    "test:watch": "ava -v -s --watch",
-    "coverage": "c8 ava -s",
     "docs": "node ../analytics-cli/bin/run docs",
     "build": "node ../../scripts/build/index.js",
     "watch": "node ../../scripts/build/_watch.js",
@@ -51,18 +48,6 @@
     "type": "git",
     "url": "git+https://github.com/DavidWells/analytics.git"
   },
-  "ava": {
-    "files": [
-      "**/**/*.test.js"
-    ],
-    "require": [
-      "esm",
-      "@babel/register"
-    ],
-    "verbose": true,
-    "failFast": true,
-    "failWithoutAssertions": false
-  },
   "dependencies": {
     "@analytics/activity-utils": "^0.1.7",
     "@analytics/localstorage-utils": "^0.1.5",
@@ -78,9 +63,6 @@
   "devDependencies": {
     "@babel/core": "^7.2.2",
     "@babel/plugin-transform-runtime": "7.5.5",
-    "@babel/preset-env": "^7.3.1",
-    "@babel/register": "^7.15.3",
-    "aws-sdk-client-mock": "^0.5.5",
-    "c8": "^7.10.0"
+    "@babel/preset-env": "^7.3.1"
   }
 }
