+++ packages/analytics-plugin-aws-pinpoint/src/pinpoint/helpers/call-aws.js
@@ -55,7 +55,8 @@ export default async function callAws(eventsRequest, config) {
   const payload = {
     body: JSON.stringify(eventsRequest),
   }
-  
+  // TODO: Not entirely sure if this is an okay way of doing this.
+  // Don't like it but couldn't think of another way in order to minimize duplicate code since server relies on sdk and browser relies on aws4fetch
   if (process.browser) {
     aws = new AwsClient(auth)
     data = await aws.fetch(endpointUrl, payload).then((d) => d.json())
